mergdata and MergeTrajectories combined, Trajectory Description and check if matching added
